requirements.txt
torch
torch-geometric
numpy
pyyaml
tqdm
bvh
scipy
